Final Project/requirements.txt
decorator==4.4.2
Django==3.0.4
django-multiselectfield==0.1.12
docutils==0.15.2
gunicorn==20.0.4
imutils==0.5.3
Jinja2==2.11.2
MarkupSafe==1.1.1
matplotlib==3.3.0
more-itertools==8.2.0
nltk==3.5
numpy==1.18.1
opencv-python==4.1.2.30
panda==0.3.1
pandas==1.0.1
pdf2image==1.13.1
Pillow==7.0.0
pipreqs==0.4.10
plac==1.1.3
pytesser3==1.0.1
pytesseract==0.3.2
regex==2020.7.14
requests==2.22.0
scikit-image==0.17.2
Textblob==0.15.3
tesseract==0.1.3
thinc==7.4.1
tifffile==2020.7.24
urllib3==1.25.8
whitenoise==5.1.0